pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nai-integrations"
version = "0.1.0"
description = "Cloud storage integrations for Django (Box, Dropbox, Google Drive, OneDrive)"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "NEMATI AI", email = "dev@nemati.ai"}
]
requires-python = ">=3.10"
dependencies = [
    "Django>=4.2",
    "django-ninja>=1.0.0",
    "requests>=2.28.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.0.0",
]
celery = [
    "celery>=5.3.0",
]
unfold = [
    "django-unfold>=0.10.0",
]

[project.urls]
Homepage = "https://github.com/NematiAI/nai-integrations"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["templates/**/*.html"]

[tool.pytest.ini_options]
python_files = ["test_*.py"]
